requirements.txt
fastapi~=0.111.0
python-dotenv
google-generativeai
torch
transformers